Move enrich with reaction summary methods to reaction service
