+++ backend/src/main/java/org/codevscovid19/stayhappyathome/controller/PostResource.java
@@ -3,7 +3,10 @@ package org.codevscovid19.stayhappyathome.controller;
 import org.codevscovid19.stayhappyathome.dto.PostDto;
 import org.codevscovid19.stayhappyathome.dto.PostReactionDto;
 import org.codevscovid19.stayhappyathome.dto.ReactionSummaryDto;
-import org.codevscovid19.stayhappyathome.entity.*;
+import org.codevscovid19.stayhappyathome.entity.Post;
+import org.codevscovid19.stayhappyathome.entity.PostReaction;
+import org.codevscovid19.stayhappyathome.entity.TargetFeeling;
+import org.codevscovid19.stayhappyathome.entity.User;
 import org.codevscovid19.stayhappyathome.login.HansNotFoundException;
 import org.codevscovid19.stayhappyathome.repository.PostReactionRepository;
 import org.codevscovid19.stayhappyathome.repository.PostRepository;
@@ -19,7 +22,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
 import java.net.URL;
-import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -55,7 +57,7 @@ public class PostResource {
   public ResponseEntity<Set<Post>> getPosts(@RequestHeader(name = USER_ID_HEADER_NAME) String userId) {
     User user = userRepository.findById(userId).orElseThrow(() -> new HansNotFoundException("User", userId));
     Set<Post> postsForUser = postService.getPostsForUser(user).stream()
-      .map(post -> enrichPostWithReactionSummary(post))
+      .map(post -> reactionService.enrichPostWithReactionSummary(post))
       .collect(Collectors.toSet());
     return ResponseEntity.ok(postsForUser);
   }
@@ -81,7 +83,7 @@ public class PostResource {
     Optional<Post> optionalPost = postRepository.findById(id);
     if (optionalPost.isPresent()) {
       Post post = optionalPost.get();
-      post = enrichPostWithReactionSummary(post);
+      post = reactionService.enrichPostWithReactionSummary(post);
       return ResponseEntity.ok(post);
     }
     return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
@@ -106,22 +108,6 @@ public class PostResource {
   public ResponseEntity<ReactionSummaryDto> getReactionSummary(@PathVariable("postId") Long postId) {
     ReactionSummaryDto reactionSummary = reactionService.getReactionSummaryForPost(postId);
     return ResponseEntity.ok(reactionSummary);
-  }
-
-  private Post enrichPostWithReactionSummary(Post post) {
-    ReactionSummaryDto postReactionSummary = reactionService.getReactionSummaryForPost(post.getId());
-    post.setReactionSummary(postReactionSummary);
-    List<Reply> replies = enrichRepliesWithReactionSummaries(post.getReplies());
-    post.setReplies(replies);
-    return post;
-  }
-
-  private List<Reply> enrichRepliesWithReactionSummaries(List<Reply> replies) {
-    for (Reply reply : replies) {
-      ReactionSummaryDto reactionSummary = reactionService.getReactionSummaryForReply(reply.getId());
-      reply.setReactionSummary(reactionSummary);
-    }
-    return replies;
   }
 
 }
